requirements.txt
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.9.14
charset-normalizer==2.1.1
decorator==4.4.2
google-api-core==2.10.1
google-api-python-client==2.62.0
google-auth==2.11.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
googleapis-common-protos==1.56.4
httplib2==0.20.4
idna==3.4
imageio==2.22.0
imageio-ffmpeg==0.4.7
moviepy==1.0.3
numpy==1.23.3
oauth2client==4.1.3
oauthlib==3.2.1
Pillow==9.2.0
proglog==0.1.10
protobuf==4.21.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-firebase==1.2.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
tqdm==4.64.1
uritemplate==4.1.1
urllib3==1.26.12
